requirements.txt
tqdm==4.66.4
bidict==0.23.1
numpy==1.26.4
pandas==1.5.3
faiss-cpu==1.8.0.post1
torch==2.6.0
transformers==4.45.2
click==8.1.7
tensorflow-datasets==4.9.6
scikit-learn==1.5.0
matplotlib==3.9.2
seaborn==0.13.2
